Add generated code for Voice (modified specs)
To generate for voicev2, use speccy to resolve the external dependencies
first

To generate sanely for voicev1, the spec was refactored to use the
components section for all objects, to give them names and create
structs that made more sense.

Check that all objects have `type: object` in the spec or OpenAPI
Generator chokes.